// accounting-management-system/gestion-comptable-frontend2/src/app/models/fournisseur.model.ts
export interface Fournisseur {
  fournisseurReference: number;
  fournisseurRaisonSocial: string;
  fournisseurAdresse: string;
  fournisseurCodePostal: string;
  fournisseurTelephone: string;
  fournisseurCodeTVA: string;
  fournisseursGouvernorat: string;
  fournisseursPays: string;
  fournisseursFax: string;
}